ethabi refactor: Function
