+++ ethabi/src/spec/mod.rs
@@ -3,7 +3,6 @@
 mod constructor;
 mod event;
 mod event_param;
-mod function;
 mod interface;
 mod operation;
 mod param;
@@ -12,7 +11,6 @@ pub mod param_type;
 pub use self::constructor::Constructor;
 pub use self::event::Event;
 pub use self::event_param::EventParam;
-pub use self::function::Function;
 pub use self::interface::Interface;
 pub use self::operation::Operation;
 pub use self::param::Param;
