+++ ethabi/src/token/token.rs
@@ -187,4 +187,50 @@ impl Token {
 			_ => None,
 		}
 	}
+
+	/// Check if all the types of the tokens match the given parameter types.
+	pub fn types_check(tokens: &[Token], param_types: &[ParamType]) -> bool {
+		param_types.len() == tokens.len() && {
+			param_types.iter().zip(tokens).all(|(param_type, token)| {
+				token.type_check(param_type)
+			})
+		}
+	}
+}
+
+
+#[cfg(test)]
+mod tests {
+	use spec::ParamType;
+	use {Token};
+
+	#[test]
+	fn test_type_check() {
+		fn assert_type_check(tokens: Vec<Token>, param_types: Vec<ParamType>) {
+			assert!(Token::types_check(&tokens, &param_types))
+		}
+
+		fn assert_not_type_check(tokens: Vec<Token>, param_types: Vec<ParamType>) {
+			assert!(!Token::types_check(&tokens, &param_types))
+		}
+
+		assert_type_check(vec![Token::Uint([0u8; 32]), Token::Bool(false)], vec![ParamType::Uint(256), ParamType::Bool]);
+		assert_type_check(vec![Token::Uint([0u8; 32]), Token::Bool(false)], vec![ParamType::Uint(32), ParamType::Bool]);
+
+		assert_not_type_check(vec![Token::Uint([0u8; 32])], vec![ParamType::Uint(32), ParamType::Bool]);
+		assert_not_type_check(vec![Token::Uint([0u8; 32]), Token::Bool(false)], vec![ParamType::Uint(32)]);
+		assert_not_type_check(vec![Token::Bool(false), Token::Uint([0u8; 32])], vec![ParamType::Uint(32), ParamType::Bool]);
+
+		assert_type_check(vec![Token::FixedBytes(vec![0, 0, 0, 0])], vec![ParamType::FixedBytes(4)]);
+		assert_not_type_check(vec![Token::FixedBytes(vec![0, 0, 0, 0])], vec![ParamType::FixedBytes(3)]);
+
+		assert_type_check(vec![Token::Array(vec![Token::Bool(false), Token::Bool(true)])], vec![ParamType::Array(Box::new(ParamType::Bool))]);
+		assert_not_type_check(vec![Token::Array(vec![Token::Bool(false), Token::Uint([0u8; 32])])], vec![ParamType::Array(Box::new(ParamType::Bool))]);
+		assert_not_type_check(vec![Token::Array(vec![Token::Bool(false), Token::Bool(true)])], vec![ParamType::Array(Box::new(ParamType::Address))]);
+
+		assert_type_check(vec![Token::FixedArray(vec![Token::Bool(false), Token::Bool(true)])], vec![ParamType::FixedArray(Box::new(ParamType::Bool), 2)]);
+		assert_not_type_check(vec![Token::FixedArray(vec![Token::Bool(false), Token::Bool(true)])], vec![ParamType::FixedArray(Box::new(ParamType::Bool), 3)]);
+		assert_not_type_check(vec![Token::FixedArray(vec![Token::Bool(false), Token::Uint([0u8; 32])])], vec![ParamType::FixedArray(Box::new(ParamType::Bool), 2)]);
+		assert_not_type_check(vec![Token::FixedArray(vec![Token::Bool(false), Token::Bool(true)])], vec![ParamType::FixedArray(Box::new(ParamType::Address), 2)]);
+	}
 }
